Modify get_ and set_state for multiple addresses
Modifies the get_state and set_state methods for TransactionContext in
the Rust SDK to take in a list of addresses to get/set instead of just
one address. This aligns with the functionality of the other SDKs. Also
modifies existing uses of SDK (smallbank_rust, inkey_rust, and xo_rust)
to call methods properly.

Signed-off-by: Logan Seeley <seeley@bitwise.io>